refactor(core): one home for the off-chip device traits; unify GpioObserver
`I2cDevice`, `SpiDevice`, `UartStreamDevice`/`UartStreamHost` and
`GpioObserver` describe the far side of a wire, not any MCU, yet each lived
inside the register model that happened to call it first. That placement is
what makes a device change an engine change, and it produced duplication
directly: `GpioObserver` was declared TWICE — byte for byte, verified against
this commit's parent — once in `peripherals::esp32::gpio` and once in
`peripherals::esp32s3::gpio`.

Two identical traits are two types. Six components that watch pad edges on
both ESP32 families each carried the same `impl` twice (`servo`, `ws2812`,
`step_dir_motor`, `h_bridge_motor`, `unipolar_stepper`, `ili9341_parallel`),
and both call sites that accept an observer — `SystemBus::install_gpio_observer`
and `AttachCtx::install_gpio_observer` — carried a two-trait bound that no
third GPIO family could satisfy without a third declaration and a third `impl`
on all six.

This moves the five traits, verbatim, into `peripherals::device` and leaves
`pub use` re-exports at every old path, so no call site outside the module
changed: `system::xtensa::esp32s3` still imports `GpioObserver` from
`esp32s3::gpio`, `crates/core/tests/` still names the S3 path, and every
intra-doc link still resolves. Twelve impls become six; two bounds become one.

Zero behaviour change. The downcast ratchet is unmoved at 193/207 — the traits
were relocated, not rewritten — so no ceiling is lowered.

The visible win is `ws2812`'s unit test: it imported
`esp32s3::gpio::{Esp32s3Gpio, GpioObserver}`, coupling the contract under test
to one silicon family. The contract now comes from `peripherals::device` and
only the concrete port model is still S3.
