+++ crates/core/src/peripherals/components/h_bridge_motor.rs
@@ -97,13 +97,7 @@ impl HBridgeMotor {
     }
 }
 
-impl crate::peripherals::esp32s3::gpio::GpioObserver for HBridgeMotor {
-    fn on_pin_change(&self, pin: u8, _from: bool, to: bool, sim_cycle: u64) {
-        self.on_gpio_edge(pin, to, sim_cycle);
-    }
-}
-
-impl crate::peripherals::esp32::gpio::GpioObserver for HBridgeMotor {
+impl crate::peripherals::device::GpioObserver for HBridgeMotor {
     fn on_pin_change(&self, pin: u8, _from: bool, to: bool, sim_cycle: u64) {
         self.on_gpio_edge(pin, to, sim_cycle);
     }
