+++ crates/core/src/peripherals/components/unipolar_stepper.rs
@@ -100,13 +100,7 @@ impl UnipolarStepper {
     }
 }
 
-impl crate::peripherals::esp32s3::gpio::GpioObserver for UnipolarStepper {
-    fn on_pin_change(&self, pin: u8, _from: bool, to: bool, sim_cycle: u64) {
-        self.on_gpio_edge(pin, to, sim_cycle);
-    }
-}
-
-impl crate::peripherals::esp32::gpio::GpioObserver for UnipolarStepper {
+impl crate::peripherals::device::GpioObserver for UnipolarStepper {
     fn on_pin_change(&self, pin: u8, _from: bool, to: bool, sim_cycle: u64) {
         self.on_gpio_edge(pin, to, sim_cycle);
     }
