+++ crates/core/src/peripherals/components/ws2812.rs
@@ -29,7 +29,7 @@
 //! (see [`crate::peripherals::esp32s3::rmt`]), which flips the routed GPIO pad
 //! through [`Esp32s3Gpio::drive_pad_output`](crate::peripherals::esp32s3::gpio::Esp32s3Gpio::drive_pad_output).
 //! This component registers as an S3
-//! [`GpioObserver`](crate::peripherals::esp32s3::gpio::GpioObserver) and decodes
+//! [`GpioObserver`](crate::peripherals::device::GpioObserver) and decodes
 //! purely from the `(pin, from, to, sim_cycle)` callbacks — accumulating each
 //! bit's HIGH duration, shifting it into a 24-bit register, and pushing a pixel
 //! every 24 bits.
@@ -201,13 +201,7 @@ impl Ws2812 {
 
 // Bridge into the ESP32-S3 GPIO observer protocol (the chip whose RMT drives the
 // pad today). `from` is unused — only the new level and the sim cycle matter.
-impl crate::peripherals::esp32s3::gpio::GpioObserver for Ws2812 {
-    fn on_pin_change(&self, pin: u8, _from: bool, to: bool, sim_cycle: u64) {
-        self.on_edge(pin, to, sim_cycle);
-    }
-}
-
-impl crate::peripherals::esp32::gpio::GpioObserver for Ws2812 {
+impl crate::peripherals::device::GpioObserver for Ws2812 {
     fn on_pin_change(&self, pin: u8, _from: bool, to: bool, sim_cycle: u64) {
         self.on_edge(pin, to, sim_cycle);
     }
@@ -431,7 +425,8 @@ mod tests {
 
     #[test]
     fn registers_as_s3_gpio_observer() {
-        use crate::peripherals::esp32s3::gpio::{Esp32s3Gpio, GpioObserver};
+        use crate::peripherals::device::GpioObserver;
+        use crate::peripherals::esp32s3::gpio::Esp32s3Gpio;
         let strip = Arc::new(Ws2812::new(PIN, 1, CPU_HZ));
         let mut g = Esp32s3Gpio::new();
         g.add_observer(strip.clone() as Arc<dyn GpioObserver>);
